+++ CHANGELOG.md
@@ -15,6 +15,7 @@ version is 3.10.
 - Add support for ntfy access tokens (#879)
 - Improve ntfy notifications (include tags, period, last ping type etc.)
 - Add an "Account closed." confirmation message after closing an account
+- Add monthly uptime percentage display in Check Details page (#773)
 
 ### Bug Fixes
 - Fix "senddeletionnotices" to recognize "Supporter" subscriptions